requirements.txt
flask==3.0.0
beautifulsoup4==4.12.2
curl_cffi>=0.7.3
psycopg2-binary==2.9.9
